# docs/api_specification.yaml
openapi: 3.0.3
info:
  title: PDF Smaller API
  description: |
    A comprehensive PDF compression service with user authentication, subscription management, 
    and bulk processing capabilities. This API provides endpoints for compressing PDF files, 
    managing user accounts, handling subscriptions, and processing multiple files in bulk.
    
    ## Features
    - Single and bulk PDF compression
    - User authentication with JWT tokens
    - Subscription management with Stripe integration
    - Usage tracking and rate limiting
    - File management and automatic cleanup
    - Comprehensive error handling and logging
    
    ## Authentication
    Most endpoints require authentication using JWT tokens. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your_jwt_token>
    ```
    
    ## Rate Limiting
    API endpoints are rate-limited based on user tier:
    - Anonymous users: 10 requests per minute
    - Free users: 20 requests per minute  
    - Premium users: 100 requests per minute
    - Pro users: 500 requests per minute
    
    ## Error Handling
    All errors follow a consistent format with error codes, messages, and request IDs for tracking.
    
  version: 1.0.0
  contact:
    name: PDF Smaller Support
    email: support@pdfsmaller.site
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.pdfsmaller.site
    description: Production server
  - url: https://staging-api.pdfsmaller.site
...
paths:
...
  /api/auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user account
      description: Create a new user account with email, password, and name
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegistration'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: User already exists
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/auth/login:
    post:
      tags:
        - Authentication
      summary: Authenticate user and get tokens
      description: Login with email and password to receive JWT tokens
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: Get a new access token using a valid refresh token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenRefresh'
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/auth/profile:
    get:
      tags:
        - Authentication
      summary: Get user profile
      description: Retrieve the authenticated user's profile information
      operationId: getUserProfile
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    put:
      tags:
        - Authentication
      summary: Update user profile
      description: Update the authenticated user's profile information
      operationId: updateUserProfile
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserProfileUpdate'
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/auth/change-password:
    post:
      tags:
        - Authentication
      summary: Change user password
      description: Change the authenticated user's password
      operationId: changePassword
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordChange'
      responses:
        '200':
          description: Password changed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # Compression Endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  schemas:
...
    UserRegistration:
      type: object
      required:
        - email
        - password
        - name
      properties:
        email:
          type: string
          format: email
          description: User's email address
          example: "user@example.com"
        password:
          type: string
          minLength: 8
          description: User's password (minimum 8 characters)
          example: "SecurePass123!"
        name:
          type: string
          minLength: 1
          maxLength: 255
          description: User's full name
          example: "John Doe"
    UserLogin:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: User's email address
          example: "user@example.com"
        password:
          type: string
          description: User's password
          example: "SecurePass123!"
    TokenRefresh:
      type: object
      required:
        - refresh_token
      properties:
        refresh_token:
          type: string
          description: Valid refresh token
          example: "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."
    AuthResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "User registered successfully"
        user:
          $ref: '#/components/schemas/User'
        tokens:
          $ref: '#/components/schemas/Tokens'
    TokenResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Token refreshed successfully"
        tokens:
          $ref: '#/components/schemas/Tokens'
    Tokens:
      type: object
      properties:
        access_token:
          type: string
          description: JWT access token
          example: "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."
        refresh_token:
          type: string
          description: JWT refresh token
          example: "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."
        expires_in:
          type: integer
          description: Access token expiration time in seconds
          example: 900
    User:
      type: object
      properties:
        id:
          type: integer
          description: User ID
          example: 1
        email:
          type: string
          format: email
          description: User's email address
          example: "user@example.com"
        name:
          type: string
          description: User's full name
          example: "John Doe"
        created_at:
          type: string
          format: date-time
          description: Account creation timestamp
          example: "2024-01-15T10:30:00Z"
        is_active:
          type: boolean
          description: Whether the account is active
          example: true
    UserProfile:
      allOf:
        - $ref: '#/components/schemas/User'
        - type: object
          properties:
            subscription:
              $ref: '#/components/schemas/Subscription'
            usage_stats:
              $ref: '#/components/schemas/UsageStats'
    UserProfileUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
          description: Updated full name
          example: "Jane Doe"
        email:
          type: string
          format: email
          description: Updated email address
          example: "jane@example.com"
    PasswordChange:
      type: object
      required:
        - current_password
        - new_password
      properties:
        current_password:
          type: string
          description: Current password
          example: "CurrentPass123!"
        new_password:
          type: string
          minLength: 8
          description: New password (minimum 8 characters)
          example: "NewPass456!"

    # Compression Schemas
...
    UsageStats:
      type: object
      properties:
        compressions_today:
          type: integer
          description: Number of compressions performed today
          example: 5
        compressions_this_month:
          type: integer
          description: Number of compressions performed this month
          example: 45
        daily_limit:
          type: integer
          description: Daily compression limit
          example: 100
        monthly_limit:
          type: integer
          description: Monthly compression limit
          example: 1000
        files_processed_today:
          type: integer
          description: Number of files processed today
          example: 8
        total_size_compressed_mb:
          type: number
          format: float
          description: Total size compressed in MB
          example: 156.7
...
    Subscription:
      type: object
      properties:
        id:
          type: integer
          description: Subscription ID
          example: 789
        user_id:
          type: integer
          description: User ID
          example: 1
        plan_id:
          type: integer
          description: Plan ID
          example: 2
        plan_name:
          type: string
          description: Plan name
          example: "premium"
        stripe_subscription_id:
          type: string
          description: Stripe subscription ID
          example: "sub_1234567890"
        status:
          type: string
          enum: [active, canceled, past_due, unpaid]
          description: Subscription status
          example: "active"
        current_period_start:
          type: string
          format: date-time
          description: Current billing period start
          example: "2024-01-01T00:00:00Z"
        current_period_end:
          type: string
          format: date-time
          description: Current billing period end
          example: "2024-02-01T00:00:00Z"
        created_at:
          type: string
          format: date-time
          description: Subscription creation timestamp
          example: "2024-01-01T00:00:00Z"
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operation completed successfully"
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              description: Error code identifier
              example: "VALIDATION_ERROR"
            message:
              type: string
              description: Human-readable error message
              example: "Invalid input provided"
            details:
              type: object
              description: Additional error details
              example: {}
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
          example: "2024-01-15T10:30:00Z"
        request_id:
          type: string
          description: Unique request identifier for tracking
          example: "req_123456789"
  responses:
    BadRequest:
      description: Bad request - invalid input or missing required fields
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized - authentication required or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    RateLimitExceeded:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
      headers:
        Retry-After:
          description: Number of seconds to wait before retrying
          schema:
            type: integer
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...